Make interactivity functional
Bind the space bar to play-pause as a proof of concept. For now,
interactive mode only works with the first player in the list of found
players, but eventually it'll have to work with all available players.
We'll likely have a callback to all the players and then have a function
that blocks and waits for input to trigger those callbacks.